pyproject.toml
[project]
name = "mobius"
version = "2.0.0"
description = "Enterprise brand governance platform with AI-powered compliance"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "langgraph>=0.2.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "supabase>=2.0.0",
    "google-generativeai>=0.3.0",
    "pdfplumber>=0.10.0",
    "pymupdf>=1.23.0",
    "pillow>=10.0.0",
    "httpx>=0.25.0",
    "structlog>=23.0.0",
    "tenacity>=8.2.0",
    "tiktoken>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "hypothesis>=6.90.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
